src/edge_server/requirements.txt
amqp==5.0.9
asgiref==3.5.0
backcall==0.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.11
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-grpc==1.0.16
django-prometheus==2.2.0
django-timezone-field==4.2.3
grpcio==1.43.0
grpcio-tools==1.43.0
importlib-metadata==4.11.1
ipdb==0.13.9
ipython==7.31.1
jedi==0.18.1
kombu==5.2.3
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
ptyprocess==0.7.0
py-grpc-prometheus==0.7.0
py-spy==0.3.11
Pygments==2.11.2
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
